Adding more user service tests, changed package
