--- src/main/java/com/example/linkshortener/security/AuthenticationResponse.java
@@ -1,17 +1,0 @@
-package com.example.linkshortener.security;
-
-public class AuthenticationResponse {
-    private String jwt;
-
-    public AuthenticationResponse(String jwt) {
-        this.jwt = jwt;
-    }
-
-    public String getJwt() {
-        return jwt;
-    }
-
-    public void setJwt(String jwt) {
-        this.jwt = jwt;
-    }
-}
