--- src/main/java/com/example/linkshortener/security/JwtAuthenticationFilter.java
@@ -1,56 +1,0 @@
-package com.example.linkshortener.security;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-@Component
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-    private final static Logger LOGGER = LoggerFactory.getLogger(JwtAuthenticationFilter.class);
-    private final JwtService jwtService;
-    private final CustomUserDetailsService myUserDetailsService;
-
-    public JwtAuthenticationFilter(JwtService jwtService, CustomUserDetailsService myUserDetailsService) {
-        this.jwtService = jwtService;
-        this.myUserDetailsService = myUserDetailsService;
-    }
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        final String authHeader = request.getHeader("Authorization");
-        final String jwt;
-        String email;
-
-        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
-            filterChain.doFilter(request, response);
-            LOGGER.info("JWT was not found");
-
-            return;
-        }
-
-        jwt = authHeader.substring(7);
-        email = jwtService.extractUsername(jwt);
-
-        if (email != null && SecurityContextHolder.getContext().getAuthentication() == null) {
-            UserDetails userDetails = myUserDetailsService.loadUserByUsername(email);
-
-            if (jwtService.isJwtValid(jwt, userDetails)) {
-                UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-                SecurityContextHolder.getContext().setAuthentication(authenticationToken);
-                LOGGER.info("User with email : " + email + " is authenticated");
-            }
-        }
-
-        filterChain.doFilter(request, response);
-    }
-}
