+++ src/main/java/com/linkshortener/security/AuthenticationResponse.java
@@ -1,0 +1,17 @@
+package com.linkshortener.security;
+
+public class AuthenticationResponse {
+    private String jwt;
+
+    public AuthenticationResponse(String jwt) {
+        this.jwt = jwt;
+    }
+
+    public String getJwt() {
+        return jwt;
+    }
+
+    public void setJwt(String jwt) {
+        this.jwt = jwt;
+    }
+}
