--- src/test/java/com/example/linkshortener/LinkShortenerApplicationTests.java
@@ -1,14 +1,0 @@
-package com.example.linkshortener;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@DataJpaTest
-class LinkShortenerApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
