--- src/test/java/com/example/linkshortener/controller/LinkControllerTest.java
@@ -1,9 +1,0 @@
-package com.example.linkshortener.controller;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class LinkControllerTest {
-
-}
