--- src/test/java/com/example/linkshortener/service/LinkServiceTest.java
@@ -1,12 +1,0 @@
-package com.example.linkshortener.service;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class LinkServiceTest {
-
-    @Test
-    void getAllLinks() {
-    }
-}
