+++ src/test/java/com/linkshortener/LinkShortenerApplicationTests.java
@@ -1,0 +1,14 @@
+package com.linkshortener;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@DataJpaTest
+class LinkShortenerApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
