+++ src/test/java/com/linkshortener/controller/LinkControllerTest.java
@@ -1,0 +1,9 @@
+package com.linkshortener.controller;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class LinkControllerTest {
+
+}
