+++ src/test/java/com/linkshortener/service/LinkServiceTest.java
@@ -1,0 +1,12 @@
+package com.linkshortener.service;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class LinkServiceTest {
+
+    @Test
+    void getAllLinks() {
+    }
+}
